aider: Implemented the `endpoint_details` function to extract and format endpoint details from the given HTML element.
# Aider chat conversation:

ASSISTANT: To implement the `endpointDetails` function as described in the TODO comment, I will create a function that extracts the method and URL from the given HTML element and formats it according to the provided example.

Here is the diff for the change:

```diff
--- app.py
+++ app.py
@@ ... @@
+def endpoint_details(element):
+    """Extract and format the endpoint details from the given HTML element."""
+    details_md = "\n### Endpoint\n\n"
+    method_element = element.find('span', class_='base__RequestMethod-sc-127j6tq-1')
+    url_element = element.find('div', class_='Text__StyledText-sc-6aor3p-0')
+    if method_element and url_element:
+        method = method_element.get_text().strip().upper()
+        url = url_element.get_text().strip()
+        details_md += f"- Method: `{method}`\n"
+        details_md += f"- Url: `{url}`\n\n"
+    return details_md
+
 # TODO: Create a generic function called endpointDetails that uses the classes of the ELEMENT to pick it out and parse out the POINTS_OF_INTEREST in this FORMAT.
 # ELEMENT
 #<div class="base__ScrollableX-sc-127j6tq-0 dFRCxS">
```

This function will be called `endpoint_details` and will take an HTML element as an argument. It will find the method and URL within the element and return a string formatted in Markdown.